move all views in views folder, create main,tickets,dashboard in templates and move all template file, fixed links in base.html, and fixed js links.
