--- project/dashboard/__init__.py
@@ -1,7 +1,0 @@
-from flask import Blueprint
-
-dashboard_blueprint = Blueprint('dashboard',
-                                __name__,
-                                template_folder='templates',
-                                static_folder='static'
-                                )
