--- project/main/__init__.py
@@ -1,6 +1,0 @@
-from flask import Blueprint
-
-homepage_blueprint = Blueprint('homepage',
-                               __name__,
-                               template_folder='templates'
-                               )
