+++ project/static/js/chart-configs.js
