+++ project/templates/base.html
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="{{ url_for('static', filename='css/main.min.css') }}" />
     <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}" />
     <link rel="stylesheet" href="{{ url_for('static', filename='css/profile.css') }}" />
-    <link rel="stylesheet" href="{{ url_for('dashboard.static', filename='style.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/dashboard.css') }}">
     <link rel="shortcut icon" href="{{ url_for('static', filename='assets/favicon.ico') }}">
     <script src="https://cdn.jsdelivr.net/npm/chart.js@3.0.2"></script>
     <script src="https://cdn.jsdelivr.net/npm/chartjs-chart-matrix@next/dist/chartjs-chart-matrix.js"></script>
