+++ project/templates/dashboard/dashboard.html
@@ -147,8 +147,8 @@
 
     </div>
 
-    <script src="{{ url_for('dashboard.static', filename='chart-configs.js') }}"></script>
-    <script src="{{ url_for('dashboard.static', filename='chart-scripts.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/chart-configs.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/chart-scripts.js') }}"></script>
     <script src="{{ url_for('static', filename='js/navbar.js') }}"></script>
 
 {% endblock %}
